refactor: shatter monolithic data_structures.h into modular headers
