+++ src/advanced_sorting_algorithms/bucket_sort.c
@@ -1,5 +1,6 @@
 #include "advanced_sorting.h"
-#include "data_structures.h"
+#include "sll.h"
+#include "array.h"
 #include "safe_input.h"
 #include "sorting_visualizer.h"
 #include <stdio.h>
