+++ src/advanced_sorting_algorithms/heap_sort.c
@@ -1,5 +1,6 @@
 #include "advanced_sorting.h"
-#include "data_structures.h" // priority_queue API (pq_init, insert, extractTop, destroy_pq)
+#include "priority_queue.h"
+#include "array.h"
 #include "safe_input.h"
 #include "sorting_visualizer.h"
 #include <stdio.h>
