+++ src/data_structures/array.h
@@ -1,0 +1,22 @@
+#ifndef ARRAY_H
+#define ARRAY_H
+
+// Swap adjacent pairs in an integer array.
+void swap_adjacent_pairs(int arr[], int length_of_array);
+
+// Find the maximum element in an integer array. Returns The maximum value in the array.
+int max_array(const int arr[], int length_of_array);
+
+// Find the minimum element in an integer array. Returns The minimum value in the array.
+int min_array(const int arr[], int length_of_array);
+
+// Print the contents of an integer array.
+void print_array(const int arr[], int length_of_array);
+
+// Reverse the contents of an integer array in place.
+void reverse_array(int arr[], int length_of_array);
+
+// Run the array demonstration module.
+void array_demo(void);
+
+#endif
