+++ src/data_structures/circular_queue.c
@@ -1,4 +1,4 @@
-#include "data_structures.h"
+#include "queue.h"
 #include "safe_input.h"
 #include <stdio.h>
 #include <stdlib.h>
