+++ src/data_structures/stack.c
@@ -1,4 +1,4 @@
-#include "data_structures.h"
+#include "stack.h"
 #include <stdbool.h>
 #include <stdio.h>
 #include <stdlib.h>
