+++ src/graph_traversals/topological_sort.c
@@ -1,4 +1,5 @@
-#include "data_structures.h"
+#include "queue.h"
+#include "stack.h"
 #include "graph_traversals.h"
 #include "returnMallocVal.h"
 #include "safe_input.h"
