+++ src/hashing/separate_chaining.c
@@ -1,4 +1,4 @@
-#include "data_structures.h"
+#include "sll.h"
 #include "hash.h"
 #include "safe_input.h"
 #include <stdio.h>
