+++ src/sorting_algorithms_n2/bubble_sort.c
@@ -1,4 +1,4 @@
-#include "data_structures.h"
+#include "array.h"
 #include "safe_input.h"
 #include "sorting_visualizer.h"
 #include "step_debugger.h"
